Rewrote the storage module in C++. The flash still does not work, but now I get the option to use RAM instead until I figure out why the flash does not work.

+++ Opentracker_3_0_1/tracker.h
@@ -6,7 +6,7 @@
 #define IMEI_LEN 15 // 14 digits + 1 check digit
 #define LED_INTERVAL 1000 // interval at which to blink status led (milliseconds)
 // millis() count for one second
-#define ONE_SEC 1000
+#define ONE_SEC ((unsigned long)1000)
 #define SECS(n) ((n)*ONE_SEC)
 #define MINS(n) ((n)*60*ONE_SEC)
 const char* BAD_IMEI = "?IMEI?";
